// src/app/models/order.model.ts
export interface OrderRecord {
  id?: string;
  lenhSanXuat: string;
  ngayXuong: string;
  ngayGiao: string;
  maHang: string;
  khachHang: string;
  tenHang: string;
  dvt: string;
  nguyenVatLieu: string;
  rong: number | string;
  dai: number | string;
  kc: number | string;
  soLuong: number;
  khoGiay: string;
  haoPhi: string | number;
}